pyproject.toml
[project]
name = "priscillafx-web"
version = "0.1.0"
description = "🌐 Priscilla FX online store and information resource written with Next.𝐣s and Fast𝓐PI. 🐘PostgreSQL / 🇷edis / 🇩ocker."
authors = [{ name = "Priscilla IT", email = "priscilla.effects@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioredis>=2.0.1",
    "asyncpg>=0.29.0",
    "celery>=5.4.0",
    "cryptography>=43.0.3",
    "fastapi-cache2[redis]>=0.2.2",
    "fastapi-users[sqlalchemy]>=13.0.0",
    "fastapi>=0.115.2",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "icecream>=2.1.3",
    "loguru>=0.7.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.36",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "alembic>=1.14.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 79
indent-width = 4

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint]
ignore = ["F403"]
